requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Data Validation
pydantic>=2.0.0

# LangChain & LLM
langchain-openai>=0.1.0
langchain-core>=0.1.0

# Database
neo4j>=5.0.0
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# Environment Variables
python-dotenv>=1.0.0

# Progress Bar
tqdm>=4.65.0
